Move TLS loader and use for S3 server as well
